backend-redes-sociales/requirements.txt
annotated-types==0.6.0
anyio==3.7.1
asgiref==3.7.2
attrs==23.1.0
autobahn==23.6.2
Automat==22.10.0
certifi==2023.7.22
cffi==1.16.0
channels==4.0.0
channels-redis==4.1.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
constantly==23.10.4
cryptography==41.0.5
daphne==4.0.0
distro==1.8.0
Django==4.2.13
django-compat==1.0.15
django-cors-headers==4.3.0
django-environ==0.11.2
django-filter==23.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-yasg==1.21.7
h11==0.14.0
httpcore==1.0.2
httpx==0.25.1
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
inflection==0.5.1
MarkupSafe==2.1.3
msgpack==1.0.7
openai==1.2.4
packaging==23.2
psycopg2-binary==2.9.9
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.5.1
pydantic_core==2.14.3
PyJWT==1.7.1
pyOpenSSL==23.3.0
pytz==2023.3.post1
PyYAML==6.0.1
redis==5.0.1
requests==2.31.0
service-identity==23.1.0
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.4
tqdm==4.66.1
txaio==23.1.1
typing_extensions==4.8.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.1.0
uvicorn==0.24.0.post1
zope.interface==6.1
